move quiz evaluation to new service
